modified project: shardingsphere-jdbc-example/other-feature-example modified information: Modify the startup class name, unified jdbc_url request parameters in the configuration file
